Distance Matrix: Extract model and view into separate module
